Alterado propriedade double para BigDecimal

+++ Mateus_Pfeffer/recharge-api/src/main/java/com/fag/domain/dto/RechargeDTO.java
@@ -1,6 +1,8 @@
 package com.fag.domain.dto;
 
-public record RechargeDTO(String id, Double value, String document, Integer operatorId, PhoneDTO phone, String receipt, Long transactionId, boolean success) {
+import java.math.BigDecimal;
+
+public record RechargeDTO(String id, BigDecimal value, String document, Integer operatorId, PhoneDTO phone, String receipt, Long transactionId, boolean success) {
 
     public RechargeDTO withReceipt(String receipt) {
         return new RechargeDTO(id(), value(), document(), operatorId(), phone(), receipt, transactionId(), success());
